+++ tests/test_custom_types.rs
@@ -6,8 +6,7 @@ use pyo3::{
     types::{PyDict, PyList, PyMapping, PySequence},
 };
 use pythonize::{
-    depythonize_bound, pythonize_custom, PythonizeDictType, PythonizeListType, PythonizeTypes,
-    Pythonizer,
+    depythonize, pythonize_custom, PythonizeDictType, PythonizeListType, PythonizeTypes, Pythonizer,
 };
 use serde::Serialize;
 use serde_json::{json, Value};
@@ -65,12 +64,10 @@ impl PythonizeTypes for PythonizeCustomList {
 fn test_custom_list() {
     Python::with_gil(|py| {
         PySequence::register::<CustomList>(py).unwrap();
-        let serialized = pythonize_custom::<PythonizeCustomList, _>(py, &json!([1, 2, 3]))
-            .unwrap()
-            .into_bound(py);
+        let serialized = pythonize_custom::<PythonizeCustomList, _>(py, &json!([1, 2, 3])).unwrap();
         assert!(serialized.is_instance_of::<CustomList>());
 
-        let deserialized: Value = depythonize_bound(serialized).unwrap();
+        let deserialized: Value = depythonize(&serialized).unwrap();
         assert_eq!(deserialized, json!([1, 2, 3]));
     })
 }
@@ -131,11 +128,10 @@ fn test_custom_dict() {
         PyMapping::register::<CustomDict>(py).unwrap();
         let serialized =
             pythonize_custom::<PythonizeCustomDict, _>(py, &json!({ "hello": 1, "world": 2 }))
-                .unwrap()
-                .into_bound(py);
+                .unwrap();
         assert!(serialized.is_instance_of::<CustomDict>());
 
-        let deserialized: Value = depythonize_bound(serialized).unwrap();
+        let deserialized: Value = depythonize(&serialized).unwrap();
         assert_eq!(deserialized, json!({ "hello": 1, "world": 2 }));
     })
 }
@@ -148,13 +144,11 @@ fn test_pythonizer_can_be_created() {
         assert!(sample
             .serialize(Pythonizer::new(py))
             .unwrap()
-            .bind(py)
             .is_instance_of::<PyDict>());
 
         assert!(sample
             .serialize(Pythonizer::custom::<PythonizeCustomDict>(py))
             .unwrap()
-            .bind(py)
             .is_instance_of::<CustomDict>());
     })
 }